+++ main.py
@@ -1279,7 +1279,7 @@ elif (mode_selection == "2" and config["general"]["modes"]["enabled"]["realtime"
         if (config["realtime"]["interface"]["display"]["speed"]["enabled"] == True and config["realtime"]["gps"]["enabled"] == True): # Display the current speed based on GPS, if enabled in the configuration.
             current_location = get_gps_location() # Get the current location.
             current_speed = convert_speed(float(current_location[2]), config["realtime"]["interface"]["display"]["speed"]["unit"]) # Convert the speed data from the GPS into the units specified by the configuration.
-            print("Current speed: " + str(current_speed) + " " + str(speed_display_unit)) # Print the current speed to the console.
+            print("Current speed: " + str(current_speed) + " " + str(config["realtime"]["interface"]["display"]["speed"]["unit"])) # Print the current speed to the console.